requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pdfkit==1.0.0
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.10.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32-ctypes==0.2.3
razorpay==1.4.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
